requirements.txt
api-display-purposes==0.0.3
asgiref==3.5.2
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
audioread==3.0.0
autopep8==1.7.0
blessings==1.7
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
clarifai==2.6.2
colorama==0.4.5
config42==0.4.2
configparser==3.8.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
dateinfer==0.2.0
decorator==4.4.2
decouple==0.0.7
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1
django-cleanup==7.0.0
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-osm-field==0.5.0
django-user-sessions==1.7.1
djangorestframework==3.13.1
easy-thumbnails==2.8.3
EasyProcess==1.1
frozenlist==1.3.3
future==0.18.2
grpcio==1.47.0
gunicorn==20.1.0
httplib2==0.21.0
huepy==1.2.1
idna==2.8
imageio==2.21.1
imageio-ffmpeg==0.4.7
iniconfig==1.1.1
itypes==1.2.0
jsonschema==2.6.0
mock==4.0.3
moviepy==1.0.3
multidict==6.0.4
mutagen==1.46.0
oauth2==1.9.0.post1
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
plyer==2.0.0
proglog==0.1.10
protobuf==3.20.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pyaes==1.6.1
pycodestyle==2.9.1
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
python-openid==2.2.5
python3-openid==3.2.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
responses==0.21.0
scipy==1.9.3
selenium==3.141.0
semantic-version==2.10.0
sentry-sdk==1.15.0
setuptools-rust==1.5.1
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.64.0
twilio==7.16.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
Wave==0.0.2
webdriverdownloader==1.1.0.3
whitenoise==6.2.0
yarl==1.8.2
cyrtranslit==1.1